Use moved time tools
