+++ willie/modules/seen.py
@@ -11,7 +11,8 @@ from __future__ import unicode_literals
 
 import time
 import datetime
-from willie.tools import Identifier, get_timezone, format_time
+from willie.tools import Identifier
+from willie.tools.time import get_timezone, format_time
 from willie.module import commands, rule, priority, thread
